Eliminate audio stuttering
Now using a float to count elapsed cpu cycles to add audio samples to get as close as possible to the output sampling rate. Also increased the size of the circular buffer to avoid underflow. This seems to have fixed the problem of write samples being dropped without introducing much delay.

+++ src/Apu.h
@@ -20,6 +20,6 @@ private:
 	bool m_evenFrame;
 	std::shared_ptr<FrameCounter> m_frameCounter;
 	std::shared_ptr<PulseChannel> m_pulse1;
-	size_t m_elapsedCpuCycles;
+	float32 m_elapsedCpuCycles;
 	std::shared_ptr<AudioDriver> m_audioDriver;
 };
